cleaned up cognito login component by extracting fetch funtions into a dedicated api
